pyproject.toml
[tool.poetry]
name = "complaintsystemproject"
version = "0.1.0"
description = ""
authors = ["Korionov Dmitriy <korionov.dmytro@onix-systems.com>"]

[tool.poetry.dependencies]
python = "^3.10"
asyncpg = "^0.26.0"
psycopg2 = "^2.9.3"
psycopg2-binary = "^2.9.3"
SQLAlchemy = "^1.4.39"
python-decouple = "^3.6"
databases = "^0.6.0"
alembic = "^1.8.1"
PyJWT = "^2.4.0"
fastapi = "^0.79.0"
passlib = "^1.7.4"
uvicorn = "^0.18.2"
bcrypt = "^3.2.2"
asyncclick = "^8.1.3"
boto3 = "^1.24.41"
requests = "^2.28.1"
sqlmodel = "^0.0.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
